misc/HtmlTestSuite/results/test-0927.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<link href="https://html.spec.whatwg.org/#attr-button-type" rel="help"/>
<link href="https://html.spec.whatwg.org/#enumerated-attribute" rel="help"/>
<meta content="button@type values are ASCII case-insensitive" name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><button type="reset">
</button><button type="ReSeT">
</button><button type="reſet">
</button><button type="submit">
</button><button type="SuBmIt">
</button><button type="ſubmit">
<script>
const button = document.querySelectorAll("button");

test(() =&gt; {
  assert_equals(button[0].type, "reset", "lowercase valid");
  assert_equals(button[1].type, "reset", "mixed case valid");
  assert_equals(button[2].type, "submit", "non-ASCII invalid");
}, "keyword reset");

test(() =&gt; {
  assert_equals(button[3].type, "submit", "lowercase valid");

  // vacuous: the invalid value default is currently submit, so even if the UA
  // treats this as invalid, the observable behaviour would still be correct
  assert_equals(button[4].type, "submit", "mixed case valid");

  // vacuous: the invalid value default is currently submit, so even if the UA
  // treats this as valid, the observable behaviour would still be correct
  assert_equals(button[5].type, "submit", "non-ASCII invalid");
}, "keyword submit");
</script>
</button></body></html>
